refs #259 added test for timezone fix + added a compile warning fix

+++ lib/QoreTimeZoneManager.cpp
@@ -186,7 +186,7 @@ QoreZoneInfo::QoreZoneInfo(QoreString &root, std::string &n_name, ExceptionSink 
                printd(1, "QoreZoneInfo::QoreZoneInfo() skipping invalid transition [%d] at %d\n", i, t.time);
                QoreDSTTransitions.erase(di);
                di = prev;
-               if (i < first_pos)
+               if ((int)i < first_pos)
                  --first_pos;
             }
          }
